services/Backend/DocumentAPI/requirements.txt
python-docx
flask>=1.1.1
gunicorn
kafka-python